requirements.txt
absl-py==1.3.0
aiosignal==1.2.0
antlr4-python3-runtime==4.9.3
astunparse==1.6.3
async-timeout==4.0.2
biopython==1.81
biotite==0.35.0
brotlipy==0.7.0
cachetools==5.2.0
certifi==2022.9.24
cffi @ file:///tmp/abs_98z5h56wf8/croots/recipe/cffi_1659598650955/work
charset-normalizer==2.1.1
contextlib2==21.6.0
cryptography @ file:///croot/cryptography_1677533068310/work
deepspeed==0.5.9
dill==0.3.5.1
DLLogger @ git+https://github.com/NVIDIA/dllogger.git@0540a43971f4a8a16693a9de9de73c1072020769
dm-tree==0.1.7
einops==0.5.0
fair-esm==2.0.0
filelock @ file:///croot/filelock_1672387128942/work
fire==0.4.0
flit_core @ file:///croot/flit-core_1679397103445/work/source/flit_core
frozenlist==1.3.1
fsspec==2022.10.0
gast==0.4.0
gmpy2 @ file:///tmp/build/80754af9/gmpy2_1645455532332/work
google-auth==2.14.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.50.0
h5py==3.7.0
hjson==3.1.0
huggingface-hub==0.10.1
idna==3.4
importlib-metadata==5.0.0
Jinja2 @ file:///croot/jinja2_1666908132255/work
joblib==1.2.0
keras==2.10.0
Keras-Preprocessing==1.1.2
libclang==14.0.6
lightning-lite==1.8.0.post1
lightning-utilities==0.3.0
Markdown==3.4.1
MarkupSafe==2.1.1
mkl-fft==1.3.1
mkl-random @ file:///tmp/build/80754af9/mkl_random_1626186064646/work
mkl-service==2.4.0
ml-collections==0.1.1
mpmath==1.2.1
msgpack==1.0.4
multidict==6.0.2
multiprocess==0.70.13
mysql-connector-python==8.0.31
networkx @ file:///croot/networkx_1678964333703/work
ninja==1.11.1
numpy==1.24.2
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
oauthlib==3.2.2
omegaconf==2.2.3
openfold @ git+https://github.com/aqlaboratory/openfold.git@4b41059694619831a7db195b7e0988fc4ff3a307
opt-einsum==3.3.0
packaging==21.3
pandas==2.0.0
Pillow==9.4.0
protobuf==3.19.6
py-cpuinfo==9.0.0
pyarrow==10.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pyOpenSSL @ file:///croot/pyopenssl_1677607685877/work
pyparsing==3.0.9
PySocks @ file:///tmp/build/80754af9/pysocks_1605305779399/work
python-dateutil==2.8.2
pytorch-lightning==1.8.0.post1
pytz==2023.3
PyYAML==6.0
regex==2022.10.31
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.2.2
scipy==1.10.1
six==1.16.0
sympy @ file:///croot/sympy_1668202399572/work
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.10.0
tensorflow-estimator==2.10.0
tensorflow-io-gcs-filesystem==0.27.0
termcolor==2.1.0
threadpoolctl==3.1.0
torch==2.0.0
torchaudio==0.8.1
torchmetrics==0.10.2
torchvision==0.15.0
tqdm==4.64.1
triton==1.0.0
typing_extensions==4.4.0
tzdata==2023.3
urllib3==1.26.12
vina==1.2.3
Werkzeug==2.2.2
wrapt==1.14.1
xxhash==3.1.0
yarl==1.8.1
zipp==3.10.0
